renaming and restructuring files
