+++ c_library/source/svt22.cc
